Remove by_channel functionality of MetricChannelResult
