+++ events-analyser/src/analysis/metrics/mod.rs
@@ -2,18 +2,17 @@ mod event_counts;
 mod false_counts;
 mod muon_lifetime;
 mod output;
+mod group_by;
 mod result;
 
 use crate::{
     engine::{FlatAlgorithm, FlatWaveform, MetricProperty},
     event::ChannelData,
 };
-use digital_muon_common::Channel;
 use serde::{Deserialize, Serialize, de::DeserializeOwned};
-use std::collections::HashMap;
 
 pub(crate) use output::MetricOutput;
-pub(crate) use result::MetricResult;
+pub(crate) use result::{PartialMetricResult, CompletedMetricResult};
 
 #[derive(Default, Clone, Debug, Serialize, Deserialize)]
 struct SumWithSumOfSqrs {
@@ -35,11 +34,15 @@ impl SumWithSumOfSqrs {
     }
 }
 
-pub(crate) trait MetricChannelResult: Clone + Serialize + DeserializeOwned {
-    type Source;
-    type Aggregrate: MetricAggregatedResult<Channel = Self>;
+pub(crate) trait MetricResultClass: Clone + Serialize + DeserializeOwned {}
 
-    fn make_default(source: Self::Source) -> Self;
+impl<T> MetricResultClass for T where T : Clone + Serialize + DeserializeOwned {}
+
+pub(crate) trait PartialMetricResultClass: MetricResultClass {
+    type Source;
+    type Complete: CompleteMetricResultClass<Partial = Self>;
+
+    fn make_default(source: &Self::Source) -> Self;
     fn push(
         &mut self,
         waveform: &FlatWaveform,
@@ -49,23 +52,9 @@ pub(crate) trait MetricChannelResult: Clone + Serialize + DeserializeOwned {
     fn len(&self) -> usize;
 }
 
-pub(crate) trait MetricAggregatedResult: Clone + Serialize + DeserializeOwned {
-    type Channel: MetricChannelResult<Aggregrate = Self>;
+pub(crate) trait CompleteMetricResultClass: MetricResultClass {
+    type Partial: PartialMetricResultClass<Complete = Self>;
 
-    fn stats_aggregator<'a, F, I>(source: I, len: f64, f: F) -> (f64, f64)
-    where
-        F: Fn(&'a Self::Channel) -> (f64, f64),
-        Self::Channel: 'a,
-        I: Iterator<Item = &'a Self::Channel> + ExactSizeIterator,
-    {
-        let (sum_of_means, sum_of_sds) = source.map(|count| f(count)).fold(
-            Default::default(),
-            |(acc_mean, acc_sd): (f64, f64), (mean, sd)| (acc_mean + mean, acc_sd + sd),
-        );
-        let mean = sum_of_means / len;
-        let sd = sum_of_sds / len;
-        (mean, sd)
-    }
-    fn aggregate(source: &HashMap<Channel, Self::Channel>) -> Self;
+    fn aggregate(source: &Self::Partial) -> Self;
     fn get_property(&self, property: &MetricProperty) -> Result<MetricOutput<f64>, String>;
 }
